Develop

+++ src/main/kotlin/com/mvp/order/domain/service/order/OrderServiceImpl.kt
@@ -9,7 +9,7 @@ import com.mvp.order.domain.model.order.enums.OrderStatusEnum
 import com.mvp.order.domain.model.product.ProductRemoveOrderDTO
 import com.mvp.order.infrastruture.repository.order.OrderProductRepository
 import com.mvp.order.infrastruture.repository.order.OrderRepository
-import com.mvp.order.domain.service.message.SnsService
+import com.mvp.order.domain.service.message.SnsAndSqsService
 import com.mvp.order.domain.service.product.ProductServiceImpl
 import com.mvp.order.domain.service.user.UserServiceImpl
 import com.mvp.order.infrastruture.entity.order.OrderEntity
@@ -20,12 +20,13 @@ import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
+import org.springframework.transaction.annotation.Transactional
 import java.math.BigDecimal
 import java.util.*
 
 @Service
 class OrderServiceImpl @Autowired constructor(
-    private val snsService: SnsService,
+    private val snsAndSqsService: SnsAndSqsService,
     private val orderRepository: OrderRepository,
     private val orderProductRepository: OrderProductRepository,
     private val orderProductResponseRepository: OrderProductResponseRepository,
@@ -77,6 +78,7 @@ class OrderServiceImpl @Autowired constructor(
         }
     }
 
+    @Transactional
     override fun createOrder(orderRequestDTO: OrderRequestDTO): OrderResponseDTO {
         var total: BigDecimal
         val listIDproduct = orderRequestDTO.orderProduct.mapNotNull { it.idProduct }.toMutableList()
@@ -96,7 +98,7 @@ class OrderServiceImpl @Autowired constructor(
             if (userDTO != null) {
                 this.idClient = userDTO.id
             }
-            this.externalId = existingOrder.externalId
+            this.externalId = existingOrder.externalId ?: UUID.randomUUID()
             this.totalPrice = BigDecimal.ZERO
             this.totalPrice = this.totalPrice.add(total)
             this.status = OrderStatusEnum.PENDING.value
@@ -106,7 +108,7 @@ class OrderServiceImpl @Autowired constructor(
         orderRequestDTO.orderProduct.forEach { it.idOrder = savedOrder.id }
         savedOrder.productList = orderProductRepository.saveAll(orderRequestDTO.toEntityList()).toMutableList()
 
-        snsService.sendMessage(mapper.writeValueAsString(OrderResponseDTO(savedOrder)))
+        snsAndSqsService.sendMessage(mapper.writeValueAsString(OrderResponseDTO(savedOrder)))
         return OrderResponseDTO(savedOrder)
     }
 
@@ -141,8 +143,14 @@ class OrderServiceImpl @Autowired constructor(
 
     override fun deleteOrderById(id: Long) {
         try {
-            orderProductRepository.deleteByIdOrder(id)
-            orderRepository.deleteById(id)
+            val orderEntityOptional = orderRepository.findByIdOrder(id)
+            if(orderEntityOptional.isPresent) {
+                val orderEntity = orderEntityOptional.get()
+                orderProductRepository.deleteByIdOrder(id)
+                orderRepository.deleteById(id)
+                orderEntity.status = OrderStatusEnum.CANCELED.value
+                snsAndSqsService.sendMessage(mapper.writeValueAsString(OrderResponseDTO(orderEntity.toDTO())))
+            }
         } catch (e: Throwable) {
             logger.error(ErrorMsgConstants.ERROR_ORDER_PRODUCT_NOT_FOUND, e.printStackTrace())
             throw Exceptions.NotFoundException(ErrorMsgConstants.ERROR_ORDER_PRODUCT_NOT_FOUND)
